Set only OMP_NUM_THREADS and reorganize threading into a utility
