+++ src/utilities/CMakeLists.txt
@@ -33,6 +33,7 @@ target_sources(safire_lib
     mpi_context.cpp
     parser.cpp
     freemem.cpp
+    threading.cpp
 )
 
 if(BUILD_UNIT_TESTS)
